Adds View number 2
Adds an sql of view number 2. This view lets you see all the rooms in a hotel with information that a client would probaly wnat to see, but the rooms are group by their capacity.
It is important to note that a Where clause should be added to check that the rooms that this view shows is available.